internal: move diagnostics infra to hir
