+++ crates/hir/src/diagnostics_sink.rs
